+++ packages/atomic/src/components/atomic-result-list/atomic-result-list.stories.tsx
@@ -1,9 +1,0 @@
-import defaultStory from '../../../.storybook/default-story';
-
-const {defaultModuleExport, exportedStory} = defaultStory(
-  'Atomic/ResultList',
-  'atomic-result-list',
-  {}
-);
-export default defaultModuleExport;
-export const DefaultResultList = exportedStory;