+++ packages/quantic/cypress/integration/document-suggestion/document.suggestion.expectations.ts
@@ -9,6 +9,8 @@ interface Fields {
   uri: string;
 }
 
+const testOrigin = 'test origin';
+
 function documentSuggestionExpectations(selector: DocumentSuggestionSelector) {
   return {
     displayAccordion: (display: boolean) => {
@@ -100,6 +102,7 @@ function documentSuggestionExpectations(selector: DocumentSuggestionSelector) {
             'documentPosition',
             index
           );
+          expect(analyticsBody.searchHub).to.eq(testOrigin);
           if (fromQuickview) {
             expect(analyticsBody.svc_action_data).to.have.property(
               'fromQuickview',
@@ -133,6 +136,7 @@ function documentSuggestionExpectations(selector: DocumentSuggestionSelector) {
             'documentPosition',
             index
           );
+          expect(analyticsBody.searchHub).to.eq(testOrigin);
         })
         .logDetail('should log the "suggestion_rate" UA event');
     },