+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.1.1...@coveo/atomic-angular@1.1.2) (2022-04-12)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.1.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.1.0...@coveo/atomic-angular@1.1.1) (2022-04-11)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+# [1.1.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.12...@coveo/atomic-angular@1.1.0) (2022-04-11)
+
+
+### Features
+
+* **atomic:** added custom styling to smart snippet answer ([#1887](https://github.com/coveo/ui-kit/issues/1887)) ([3932de3](https://github.com/coveo/ui-kit/commit/3932de3d10b67f539b9aa2d5f7d8a4e1cc08c2b4))
+
+
+
+
+
+## [0.8.12](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.11...@coveo/atomic-angular@0.8.12) (2022-04-07)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [0.8.11](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.10...@coveo/atomic-angular@0.8.11) (2022-04-06)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [0.8.10](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.9...@coveo/atomic-angular@0.8.10) (2022-04-05)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [0.8.9](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.8...@coveo/atomic-angular@0.8.9) (2022-04-05)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
 ## [0.8.8](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.7...@coveo/atomic-angular@0.8.8) (2022-04-04)
 
 **Note:** Version bump only for package @coveo/atomic-angular