+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.51.23](https://github.com/coveo/ui-kit/compare/headless-react@0.51.22...headless-react@0.51.23) (2022-04-12)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.51.22](https://github.com/coveo/ui-kit/compare/headless-react@0.51.21...headless-react@0.51.22) (2022-04-11)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.51.21](https://github.com/coveo/ui-kit/compare/headless-react@0.51.20...headless-react@0.51.21) (2022-04-11)
+
+
+### Bug Fixes
+
+* **atomic:** fix atomic react rendering function ([#1909](https://github.com/coveo/ui-kit/issues/1909)) ([28217bf](https://github.com/coveo/ui-kit/commit/28217bfc4b1dc130592cbc7749ee76c4ce437bce))
+
+
+
+
+
+## [0.51.20](https://github.com/coveo/ui-kit/compare/headless-react@0.51.19...headless-react@0.51.20) (2022-04-07)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.51.19](https://github.com/coveo/ui-kit/compare/headless-react@0.51.18...headless-react@0.51.19) (2022-04-06)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.51.18](https://github.com/coveo/ui-kit/compare/headless-react@0.51.17...headless-react@0.51.18) (2022-04-06)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.51.17](https://github.com/coveo/ui-kit/compare/headless-react@0.51.16...headless-react@0.51.17) (2022-04-05)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.51.16](https://github.com/coveo/ui-kit/compare/headless-react@0.51.15...headless-react@0.51.16) (2022-04-05)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 ## [0.51.15](https://github.com/coveo/ui-kit/compare/headless-react@0.51.14...headless-react@0.51.15) (2022-04-05)
 
 **Note:** Version bump only for package headless-react